updated error handling

+++ YSqliteStatement/YSqliteStatement.h
@@ -89,7 +89,7 @@ typedef enum YSqliteStatmentStatus
 - (BOOL)bindText:(NSString *)value key:(NSString *)key;
 - (BOOL)bindValue:(id)value key:(NSString *)key;
 //- (BOOL)bindValuesAndKeysAndTypes:(id)firstValue, ...;
-- (BOOL)bindValue:(id)value index:(NSUInteger)index;
+- (BOOL)bindValue:(id)value index:(int)index;
 - (BOOL)bindDate:(NSDate *)value index:(int)index;
 - (BOOL)bindDate:(NSDate *)value key:(NSString *)key;
 - (BOOL)bindNullIndex:(int)index;
